Naprawione: działające drzewa oraz ich dodawanie na mape, Nie działa: timer po gameover, Zmiany: wysokość gracza jest teraz 3/4 tekstury, renderowanie obiektów i gracza następuje w Map.cpp

+++ SDL3-RoughLite/Enemy.hpp
@@ -51,6 +51,6 @@ private:
     int currentFrame; // Aktualna klatka
     int currentRow; // Aktualny wiersz
     int totalFrames; // Ca³kowita liczba klatek
-    Uint32 lastFrameTime; // Ostatni czas klatki
-    Uint32 frameDuration; // Czas trwania klatki
+    Uint64 lastFrameTime; // Ostatni czas klatki
+    Uint64 frameDuration; // Czas trwania klatki
 };
