+++ SDL3-RoughLite/RangeRover.hpp
@@ -52,8 +52,8 @@ private:
 	int currentFrame; // Aktualna klatka
 	int currentRow; // Aktualny wiersz
 	int totalFrames; // Ca³kowita liczba klatek
-	Uint32 lastFrameTime; // Ostatni czas klatki
-	Uint32 frameDuration; // Czas trwania klatki
+	Uint64 lastFrameTime; // Ostatni czas klatki
+	Uint64 frameDuration; // Czas trwania klatki
 
 	float angle; // K¹t
 	float radius; // Promieñ
